Expanded README to be a bit clearer.
RH at the OpenSMC meeting had Javascript disabled, so that's called
out now.

Made use of python or http-serve web servers slightly clearer and
removed the prompts that could cause confusion.

Explained what happens if all for columns for a setting are
given, and set.

Added a few more examples, and a bit of motivation around numeric
formatters.